requirements.txt
psycopg2>=2.8.4,<2.9.0
pyTelegramBotAPI>=3.6.6,<3.7.0
python-dotenv>=0.10.3,<0.11.0
